Fix vdeslirp_send FFI return type: ssize_t, not c_int
The C library returns ssize_t (i64 on 64-bit), not int. Using c_int
would truncate return values on ABI mismatch.

+++ tt-bh-linux-rs/src/slirp_ffi.rs
@@ -39,7 +39,7 @@ extern "C" {
         slirp: *mut VdeSlirp,
         buf: *const u8,
         len: libc::size_t,
-    ) -> libc::c_int;
+    ) -> libc::ssize_t;
 
     pub fn inet_aton(cp: *const libc::c_char, inp: *mut InAddr) -> libc::c_int;
 }
